+++ frontend/src/components/Calendar/Calendar.tsx
@@ -1,163 +1,14 @@
 import React, { useState } from 'react';
-import { Box, Button, Grid, Stack } from '@mui/material';
+import { Grid } from '@mui/material';
 import './Calendar.scss';
 import { EventProps } from 'pages/Props/Event';
-import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos';
-import ArrowBackIosNewIcon from '@mui/icons-material/ArrowBackIosNew';
-import { CalendarMonth } from '@mui/icons-material';
-
-interface EventDataProps {
-  key: number; // The index of the event which this object refects to in the events list.
-  beginDate: Date; // The start date of the event.
-  endDate: Date; // The end date of the event.
-  blocked: boolean; // Wheter take part of a chain of simultaneous events which use all the horizontal space of a Day component.
-  size: number; // Relative horizontal size of the event.
-  sameTimeEvent: Array<number>; // List of simultaneous events.
-  coupleEvents: Array<Array<number>>; // List of simultaneous events which have all a common area of time.
-}
-
-/**
- * The EventBlock component, which is an event in the calendar display.
- * @param day The number of the day in the week.
- * @param event The event object with the event data.
- * @returns The button to display, already placed and resized.
- */
-function EventBlock(props: { day: number; event: EventProps }) {
-  const { day, event } = props;
-  const beginDate = new Date(event.date);
-  const endDate = new Date(event.end_date);
-  let todayBegin = false;
-  let startTime = 24;
-
-  // Set the time when the event begins in the day.
-  if (beginDate.getDay() === day % 7) {
-    todayBegin = true;
-    if (
-      beginDate.getHours() !== 0 ||
-      beginDate.getMinutes() !== 0 ||
-      beginDate.getSeconds() !== 0
-    ) {
-      startTime =
-        23 -
-        beginDate.getHours() +
-        (59 - beginDate.getMinutes()) / 60 +
-        (60 - beginDate.getSeconds()) / 3600;
-    }
-  }
-
-  // Set the duration of the event.
-  let duration;
-  console.log(day, endDate);
-  if (todayBegin) {
-    duration = (endDate.getTime() - beginDate.getTime()) / 3600000;
-  } else if (endDate.getDay() === day % 7) {
-    duration =
-      endDate.getHours() +
-      endDate.getMinutes() / 60 +
-      endDate.getSeconds() / 3600;
-  } else {
-    duration = 24;
-  }
-  return (
-    <Button
-      variant="contained"
-      style={{
-        position: 'absolute',
-        minWidth: `1px`,
-        // width: `${(120 * event.effectiveSize) / event.globalSize}px`,
-        height: `${Math.min(duration, startTime) * 20}px`,
-        padding: '0px',
-        transform: `translate(${
-          -60 + (120 * event.position) / event.globalSize
-        }px, -${startTime * 20}px)`,
-      }}
-    >
-      {event.title}
-    </Button>
-  );
-}
-
-/**
- * Function that creates the TimeBlock component, which are the backround lines in the calendar.
- * @param startTime The hour value, to set where and how the component has to be placed.
- * @returns The TimeBlock component.
- */
-function TimeBlock(props: { startTime: number }) {
-  const { startTime } = props;
-
-  if (startTime % 2 === 0) {
-    return <Box className="TimeBlockEven"></Box>;
-  }
-
-  return <Box className="TimeBlockOdd"></Box>;
-}
-
-/**
- * Function that creates the DayInfos component, which display the hours in the day.
- * @return The DayInfos component.
- */
-function DayInfos() {
-  return (
-    <div style={{ display: 'block' }}>
-      <p>0</p>
-      <p>2</p>
-      <p>4</p>
-      <p>6</p>
-      <p>8</p>
-      <p>10</p>
-      <p>12</p>
-      <p>14</p>
-      <p>16</p>
-      <p>18</p>
-      <p>20</p>
-      <p>22</p>
-    </div>
-  );
-}
-
-/**
- * The Day component which contains all the TimeBlocks and EventBlocks of this day.
- * @param dayValue The value of the day in the week.
- * @param day The day in the week.
- * @param events The list of events in corresponding day.
- * @returns The Day component.
- */
-function Day(props: {
-  dayValue: number;
-  day: string;
-  events: Array<EventProps>;
-}) {
-  const { dayValue, day, events } = props;
-
-  const dayChain = [];
-  for (let hour = 0; hour < 24; hour++) {
-    dayChain.push(<TimeBlock key={hour} startTime={hour}></TimeBlock>);
-  }
-
-  return (
-    <div id={day} style={{ display: 'block' }}>
-      {day}
-      {dayChain}
-      {events.map((event) => (
-        <EventBlock key={event.slug} day={dayValue} event={event} />
-      ))}
-    </div>
-  );
-}
-
-/**
- * Function that check if a date is between two others.
- * @param eventDate The date to check.
- * @param beginDate The minimal date.
- * @param endDate The maximal date.
- * @returns The boolean that tells whether the date is between the others
- */
-function betweenDate(eventDate: Date, beginDate: Date, endDate: Date) {
-  if (beginDate <= eventDate && eventDate < endDate) {
-    return true;
-  }
-  return false;
-}
+import { betweenDate } from '../../utils/date';
+import { ppcm } from '../../utils/maths';
+import { isInArray } from '../../utils/array';
+import { Day } from './Day/Day';
+import { EventDataProps } from './EventDataProps/EventDataProps';
+import { DayInfos } from './DayInfos/DayInfos';
+import { ChooseWeek } from './ChooseWeek/ChooseWeek';
 
 /**
  * Function that sort event date-wise.
@@ -173,7 +24,6 @@ function getEventWithDate(
 ) {
   const sortedEvents = new Array<EventProps>();
   events.forEach((event) => {
-    console.log(event.date);
     // Sort with end date too.
     if (
       betweenDate(new Date(event.date), beginDate, endDate) ||
@@ -210,40 +60,6 @@ function sameTime(event: EventDataProps, event2Compare: EventDataProps) {
 }
 
 /**
- * A function that returns the pgcd between 2 numbers.
- * @param a The first number.
- * @param b The second number.
- * @returns The pgcd of the 2 numbers.
- */
-function pgcd(a: number, b: number) {
-  let dividende = a;
-  let reste = b;
-  let temp;
-  const maxIter = 1000;
-  let iter = 0;
-  while (reste !== 0 && iter < maxIter) {
-    temp = dividende % reste;
-    dividende = reste;
-    reste = temp;
-    iter += 1;
-  }
-  if (reste !== 0) {
-    console.warn('Un pgcd a échoué');
-  }
-  return dividende;
-}
-
-/**
- * A function that returns the ppcm between 2 numbers.
- * @param a The first number.
- * @param b The second number.
- * @returns The ppcm of the 2 numbers.
- */
-function ppcm(a: number, b: number) {
-  return (a * b) / pgcd(a, b);
-}
-
-/**
  * Function that returns if an event have a common time area with a list of events that have a common time area.
  * @param key The key of the event to check.
  * @param eventsList The list of key of the events which with the event has to be checked.
@@ -265,22 +81,6 @@ function allSameTime(
 }
 
 /**
- * Function that returns if all the value of an array are in another one.
- * @param ar1 The array to check if the values are in the other.
- * @param ar2 The array to check if it contains the values.
- * @returns If the values of the first array are in the second.
- */
-function isInArray(ar1: Array<number>, ar2: Array<number>) {
-  let same = true;
-  let iterator = 0;
-  while (same && iterator < ar1.length) {
-    same = ar2.includes(ar1[iterator]);
-    iterator += 1;
-  }
-  return same;
-}
-
-/**
  * Get couples of events with maximal events foreach event and set the size of each event with it.
  * @param events List of events.
  * @param eventsData List of events data.
@@ -294,7 +94,7 @@ function blockedChains(
     maxSize: 0,
     blockedChains: [],
   };
-  let previousMaxSize;
+  let previousMaxSize: number;
   const blockedEventsChain = [];
   let globalSize = 1;
   eventsData.forEach((eventData) => {
@@ -342,7 +142,7 @@ function blockedChains(
  * @param eventsData List of events data.
  */
 function createCoupleEvents(eventsData: Array<EventDataProps>) {
-  let coupleEventsLength;
+  let coupleEventsLength: number;
   eventsData.forEach((eventData) => {
     for (let j = 0; j < eventData.sameTimeEvent.length; j++) {
       coupleEventsLength = eventData.coupleEvents.length;
@@ -369,9 +169,9 @@ function eventSizeReajust(
   events: Array<EventProps>,
   eventsData: Array<EventDataProps>
 ) {
-  let event2reajust;
-  let size2Add;
-  let sizeUsed;
+  let event2reajust: Array<number>;
+  let size2Add: number;
+  let sizeUsed: number;
   blockedEventsChain.forEach((eventChain) => {
     sizeUsed = 0;
     event2reajust = [];
@@ -475,7 +275,7 @@ function placeEvents(
   let select = 0;
   let chainPlaced = false;
   let allPlaced = false;
-  let newValues;
+  let newValues: { newSelect: number; chainPlaced: boolean };
   while (!allPlaced) {
     while (!chainPlaced && select < eventsBlockedChain[0].length) {
       newValues = placeChainEvent(
@@ -541,7 +341,7 @@ function setSameTimeEvents(events: Array<EventProps>) {
   const blockedEventsChain = blockedChains(events, eventsData);
 
   // Set the blocked attribute to true for all events that take part of a couple chain that takes all horizontal space.
-  let sizeUsed;
+  let sizeUsed: number;
   blockedEventsChain.forEach((eventChain) => {
     sizeUsed = 0;
     eventChain.forEach((eventKey) => {
@@ -563,6 +363,12 @@ function setSameTimeEvents(events: Array<EventProps>) {
   }
 }
 
+/**
+ *
+ * @param event
+ * @param mondayDate
+ * @param sortEvents
+ */
 function isInDay(event, mondayDate, sortEvents) {
   const checkBeginDate = new Date(
     mondayDate.getFullYear(),
@@ -612,136 +418,9 @@ function sortInWeek(oldSortEvents: Array<EventProps>, mondayDate) {
     isInDay(event, mondayDate, sortEvents);
   });
   sortEvents.forEach((eventsList) => {
-    console.log(eventsList);
     setSameTimeEvents(eventsList);
   });
   return sortEvents;
-}
-
-function DateBox(props: { date: Date; endDate: Date }) {
-  const { date, endDate } = props;
-  const sunday = new Date(
-    endDate.getFullYear(),
-    endDate.getMonth(),
-    endDate.getDate()
-  );
-  sunday.setDate(sunday.getDate() - 1);
-  return (
-    <div>
-      <Button
-        variant="outlined"
-        onClick={() => {
-          console.log('tap');
-        }}
-      >
-        {`${date.toLocaleDateString('fr-FR', {
-          weekday: 'long',
-          year: 'numeric',
-          month: 'long',
-          day: 'numeric',
-        })} au `}
-        {sunday.toLocaleDateString('fr-FR', {
-          weekday: 'long',
-          year: 'numeric',
-          month: 'long',
-          day: 'numeric',
-        })}
-        <CalendarMonth></CalendarMonth>
-      </Button>
-    </div>
-  );
-}
-
-function ChangeWeek(props: {
-  action: 'previous' | 'next';
-  beginDate: Date;
-  endDate: Date;
-  updateBegin: any;
-  updateEnd: any;
-}) {
-  const { action, beginDate, endDate, updateBegin, updateEnd } = props;
-  if (action === 'previous') {
-    return (
-      <Button
-        variant="outlined"
-        onClick={() => {
-          const newBeginDate = new Date(
-            beginDate.getFullYear(),
-            beginDate.getMonth(),
-            beginDate.getDate()
-          );
-          const newEndDate = new Date(
-            endDate.getFullYear(),
-            endDate.getMonth(),
-            endDate.getDate()
-          );
-          newBeginDate.setDate(newBeginDate.getDate() - 7);
-          updateBegin(newBeginDate);
-          newEndDate.setDate(endDate.getDate() - 7);
-          updateEnd(newEndDate);
-        }}
-      >
-        <ArrowBackIosNewIcon></ArrowBackIosNewIcon>
-      </Button>
-    );
-  }
-  if (action === 'next') {
-    return (
-      <Button
-        variant="outlined"
-        onClick={() => {
-          const newBeginDate = new Date(
-            beginDate.getFullYear(),
-            beginDate.getMonth(),
-            beginDate.getDate()
-          );
-          const newEndDate = new Date(
-            endDate.getFullYear(),
-            endDate.getMonth(),
-            endDate.getDate()
-          );
-          newBeginDate.setDate(newBeginDate.getDate() + 7);
-          updateBegin(newBeginDate);
-          newEndDate.setDate(endDate.getDate() + 7);
-          updateEnd(newEndDate);
-        }}
-      >
-        <ArrowForwardIosIcon></ArrowForwardIosIcon>
-      </Button>
-    );
-  }
-  console.warn('Appel de ChangeWeek mal effectué');
-  return <p>Error</p>;
-}
-
-function ChooseWeek(props: {
-  beginDate: Date;
-  endDate: Date;
-  updateBegin: any;
-  updateEnd: any;
-}) {
-  const { beginDate, endDate, updateBegin, updateEnd } = props;
-  return (
-    <div id="day" style={{ display: 'flex' }}>
-      <ChangeWeek
-        action="previous"
-        beginDate={beginDate}
-        endDate={endDate}
-        updateBegin={updateBegin}
-        updateEnd={updateEnd}
-      ></ChangeWeek>
-      <Stack spacing={3}>
-        <DateBox date={beginDate} endDate={endDate}></DateBox>
-      </Stack>
-      <ChangeWeek
-        action="next"
-        beginDate={beginDate}
-        endDate={endDate}
-        updateBegin={updateBegin}
-        updateEnd={updateEnd}
-      ></ChangeWeek>
-    </div>
-  );
 }
 
 /**
@@ -753,9 +432,6 @@ function Calendar(props: { events: Array<EventProps> }) {
   const { events } = props;
 
   console.log(events);
-
-  // Selection day to implement
-  console.log('implémenter la sélection des jours');
 
   const tempMondayOfTheWeek = new Date();
   tempMondayOfTheWeek.setDate(
@@ -785,7 +461,11 @@ function Calendar(props: { events: Array<EventProps> }) {
   );
 
   console.log(beginOfWeek, endOfWeek);
-  let sortEvents = getEventWithDate(events, beginOfWeek, endOfWeek);
+  const sortEvents: Array<EventProps> = getEventWithDate(
+    events,
+    beginOfWeek,
+    endOfWeek
+  );
 
   // Delete when end_date added
   console.log('endDate à virer');
@@ -796,7 +476,7 @@ function Calendar(props: { events: Array<EventProps> }) {
   });
   //
 
-  sortEvents = sortInWeek(sortEvents, beginOfWeek);
+  const newSortEvents = sortInWeek(sortEvents, beginOfWeek);
   return (
     <>
       <p>Le calendrier</p>
@@ -808,8 +488,8 @@ function Calendar(props: { events: Array<EventProps> }) {
         updateEnd={setEndOfWeek}
       ></ChooseWeek>
       <div id="Calendar" style={{ display: 'flex' }}>
-        <Grid container spacing={0} md={12}>
-          <Grid item md={1}>
+        <Grid container spacing={0} xs={12}>
+          <Grid item xs={1}>
             <DayInfos />
           </Grid>
           {[
@@ -822,12 +502,12 @@ function Calendar(props: { events: Array<EventProps> }) {
             'Dimanche',
           ].map((day, number) => {
             return (
-              <Grid item md={1} key={day}>
+              <Grid item xs={1.5} key={day}>
                 <Day
                   key={day}
                   dayValue={number + 1}
                   day={day}
-                  events={sortEvents[number]}
+                  events={newSortEvents[number]}
                 />
               </Grid>
             );
