+++ frontend/src/components/Calendar/DayInfos/DayInfos.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+/**
+ * Function that creates the DayInfos component, which display the hours in the day.
+ * @return The DayInfos component.
+ */
+export function DayInfos() {
+  return (
+    <div style={{ display: 'block' }}>
+      <p>0</p>
+      <p>2</p>
+      <p>4</p>
+      <p>6</p>
+      <p>8</p>
+      <p>10</p>
+      <p>12</p>
+      <p>14</p>
+      <p>16</p>
+      <p>18</p>
+      <p>20</p>
+      <p>22</p>
+    </div>
+  );
+}
