+++ frontend/src/utils/date.ts
@@ -35,3 +35,17 @@ export function formatTime(
 ) {
   return formatDate(date, undefined, timeFormat);
 }
+
+/**
+ * Function that check if a date is between two others.
+ * @param eventDate The date to check.
+ * @param beginDate The minimal date.
+ * @param endDate The maximal date.
+ * @returns The boolean that tells whether the date is between the others
+ */
+export function betweenDate(eventDate: Date, beginDate: Date, endDate: Date) {
+  if (beginDate <= eventDate && eventDate < endDate) {
+    return true;
+  }
+  return false;
+}
